Change the PC angle command to be of type float

+++ application/infantry_cmd.h
@@ -106,8 +106,8 @@ struct cmd_gimbal_angle
         uint8_t pitch_mode:1;
     }bit;
   } ctrl;
-  int16_t pitch;
-  int16_t yaw;
+  float pitch;
+  float yaw;
 };
 
 struct cmd_chassis_speed
